MAINT: Rename of verification utils module
