+++ src/porepy/applications/verification_setups/manu_flow_comp_frac.py
@@ -20,7 +20,7 @@ import numpy as np
 import sympy as sym
 
 import porepy as pp
-from porepy.applications.verification_setups.verifications_utils import (
+from porepy.applications.verification_setups.verification_utils import (
     VerificationUtils,
 )
